+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4040,7 +4040,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KX634Z30WVHVFFX50CTQG7XV"),
+  id: z.string().default("01KX67H3B0V3BRJX5PVHTNAQGE"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),