+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-10T13:22:57.648Z",
+    "2026-07-10T14:39:35.652Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {