+++ packages/orq-rc/RELEASES.md
@@ -10699,3 +10699,13 @@ Based on:
 - [typescript v4.12.0-rc.14] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.14 - packages/orq-rc
+
+## 2026-07-10 14:42:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.2 (2.916.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.15] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.15 - packages/orq-rc